+++ l10n/app_localizations.dart
@@ -101,7 +101,7 @@ abstract class AppLocalizations {
   /// No description provided for @appTitle.
   ///
   /// In en, this message translates to:
-  /// **'oshi@log'**
+  /// **'Oshi@log'**
   String get appTitle;
 
   /// No description provided for @home.